refactor: move k6 test scenarios to separate files
