+++ app/tests/perf/scenarios/uploads_small.js
@@ -1,0 +1,85 @@
+/* eslint-disable */
+import http from 'k6/http';
+import exec from 'k6/execution';
+import { FormData } from 'https://jslib.k6.io/formdata/0.0.2/index.js';
+
+/*
+    Upload test files generated by running make generate_perf_files` in project root
+    or using command:
+          head -c [file size] </dev/urandom >[file name]
+
+          example:
+          head -c 100000 </dev/urandom >file_1M.bin
+ */
+const file100KB = open('../../../../k6files/file_100KB.bin', 'b');
+const file1M = open('../../../../k6files/file_1M.bin', 'b');
+
+const createAttachmentPayload = JSON.stringify({
+  id: 'createAttachmentMutation',
+  query: `
+      mutation createAttachmentMutation($input: CreateAttachmentInput!) {
+        createAttachment(input: $input) {
+          attachment {
+            rowId
+            file
+          }
+        }
+      }`,
+  variables: {
+    input: {
+      attachment: {
+        file: null,
+        fileName: 'file.bin',
+        fileSize: '1MB',
+        fileType: 'test',
+        applicationId: 1,
+      },
+    },
+  },
+});
+
+const createFormData = (file) => {
+  const fd = new FormData();
+  fd.append('operations', createAttachmentPayload);
+
+  fd.append('map', '{"nfile":["variables.input.attachment.file"]}');
+  fd.append('nfile', file);
+
+  return fd;
+};
+
+const createPostParams = (formDataBoundary) => {
+  const postParams = {
+    cookies: {
+      'mocks.auth_role': 'ccbc_auth_user',
+      'mocks.mocked_timestamp': 1660924800,
+    },
+    headers: {
+      'Content-Type': `multipart/form-data; boundary=${formDataBoundary}`,
+    },
+  };
+
+  return postParams;
+};
+const fd_100KB = createFormData(http.file(file100KB, 'file_10M.bin'));
+const fd_1MB = createFormData(http.file(file1M, 'file_1M.bin'));
+
+const uploads = () => {
+  http.post(
+    __ENV.APP_HOST + '/graphql',
+    fd_1MB.body(),
+    Object.assign({}, createPostParams(fd_1MB.boundary), {
+      tags: { name: 'createAttachmentMutation' },
+    })
+  );
+
+  http.post(
+    __ENV.APP_HOST + '/graphql',
+    fd_100KB.body(),
+    Object.assign({}, createPostParams(fd_100KB.boundary), {
+      tags: { name: 'createAttachmentMutation' },
+    })
+  );
+};
+
+export default uploads;
